requirements.txt
alabaster==0.7.12
awscli==1.23.9
Babel==2.10.1
botocore==1.25.9
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
coverage==6.3.2
docutils==0.15.2
flake8==4.0.1
idna==3.3
imagesize==1.3.0
importlib-metadata==4.2.0
Jinja2==3.1.2
jmespath==1.0.0
joblib==1.1.0
MarkupSafe==2.1.1
mccabe==0.6.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
numexpr==2.8.1
numpy==1.21.6
packaging==21.3
pandas==1.4.2
pyasn1==0.4.8
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.12.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==5.4.1
requests==2.27.1
rsa==4.7.2
s3transfer==0.5.2
scikit-surprise==1.1.1
scipy==1.7.3
six==1.16.0
snowballstemmer==2.2.0
Sphinx==4.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
typing_extensions==4.2.0
urllib3==1.26.9
wincertstore==0.2
zipp==3.8.0
